Rework ja_keyboard.pyw completely
This branch does away with the word listeners and instead keeps track of a working string according to what is typed. This has the advantage of translating immediately after a mora is matched. Kanji is still supported with mostly the same process as the original branch. 

There is an issue where very quick or simultaneous inputs are not processed correctly, which I tried to alleviate a bit by using time's sleep function, so there is another import. This band-aid is mostly optional, as this issue isn't really present when typing normally, but holding down keys, for example, fails to translate properly.